+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 5.2.5 (11/21/2017 - )
 * test: replace hamcrest with assertj
+* http: in local dev env, allow developer to run multiple apps on different port, either put -Dsys.http.port=8080 in Intellij Run configuration, or ./gradlew -Dsys.http.port=8080 :some-service:run
 
 ### 5.2.4 (11/20/2017 - 11/20/2017)
 * test: update junit to 5.0.2